face-auth-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
face-recognition==1.3.0
numpy==1.24.3
python-dotenv==1.0.0
Pillow==10.0.0
gunicorn==21.2.0 